thisjava 수정

+++ ThisJava/src/ch06/sec08/exam01/Calculator.java
@@ -21,7 +21,7 @@ public class Calculator {
 	
 	//호출 시 두 정수 값을 전달받고,
 		//호출한 곳으로 결과값 double을 리턴하는 메소드 선언
-	int plus(double x, double y) {
+	double plus(double x, double y) {
 		double result = (double)x / (double)y;
 		return result;
 		}
